Goodbye patch, hello render! 👋
Rename patch→render and use super simple signature.

There is no longer a way to patch a specific element and recycling
has been removed. 100% client-side apps.

The render function is essentially an overly simplified ReactDOM.render,
without component lifecycle, based on Hyperapp's virtual DOM.

Mounting takes over the supplied container so make sure your container is
empty.
