+++ test/h.test.js
@@ -2,7 +2,7 @@ import { h } from "../src"
 
 test("empty vnode", () => {
   expect(h("div")).toEqual({
-    nodeName: "div",
+    name: "div",
     attributes: {},
     children: []
   })
@@ -10,13 +10,13 @@ test("empty vnode", () => {
 
 test("vnode with a single child", () => {
   expect(h("div", {}, ["foo"])).toEqual({
-    nodeName: "div",
+    name: "div",
     attributes: {},
     children: ["foo"]
   })
 
   expect(h("div", {}, "foo")).toEqual({
-    nodeName: "div",
+    name: "div",
     attributes: {},
     children: ["foo"]
   })
@@ -24,24 +24,24 @@ test("vnode with a single child", () => {
 
 test("positional String/Number children", () => {
   expect(h("div", {}, "foo", "bar", "baz")).toEqual({
-    nodeName: "div",
+    name: "div",
     attributes: {},
     children: ["foo", "bar", "baz"]
   })
 
   expect(h("div", {}, 0, "foo", 1, "baz", 2)).toEqual({
-    nodeName: "div",
+    name: "div",
     attributes: {},
     children: [0, "foo", 1, "baz", 2]
   })
 
   expect(h("div", {}, "foo", h("div", {}, "bar"), "baz", "quux")).toEqual({
-    nodeName: "div",
+    name: "div",
     attributes: {},
     children: [
       "foo",
       {
-        nodeName: "div",
+        name: "div",
         attributes: {},
         children: ["bar"]
       },
@@ -61,7 +61,7 @@ test("vnode with attributes", () => {
   }
 
   expect(h("div", attributes, "baz")).toEqual({
-    nodeName: "div",
+    name: "div",
     attributes,
     children: ["baz"]
   })
@@ -69,7 +69,7 @@ test("vnode with attributes", () => {
 
 test("skip null and Boolean children", () => {
   const expected = {
-    nodeName: "div",
+    name: "div",
     attributes: {},
     children: []
   }
@@ -77,37 +77,4 @@ test("skip null and Boolean children", () => {
   expect(h("div", {}, true)).toEqual(expected)
   expect(h("div", {}, false)).toEqual(expected)
   expect(h("div", {}, null)).toEqual(expected)
-})
-
-test("components", () => {
-  const Component = (attributes, children) => h("div", attributes, children)
-
-  expect(h(Component, { id: "foo" }, "bar")).toEqual({
-    nodeName: "div",
-    attributes: { id: "foo" },
-    children: ["bar"]
-  })
-
-  expect(h(Component, { id: "foo" }, [h(Component, { id: "bar" })])).toEqual({
-    nodeName: "div",
-    attributes: { id: "foo" },
-    children: [
-      {
-        nodeName: "div",
-        attributes: { id: "bar" },
-        children: []
-      }
-    ]
-  })
-})
-
-test("component with no attributes adds default attributes", () => {
-  const Component = ({ name = "world" }, children) =>
-    h("div", {}, "Hello " + name)
-
-  expect(h(Component)).toEqual({
-    nodeName: "div",
-    attributes: {},
-    children: ["Hello world"]
-  })
 })
